uninitialized affecting output?
